2025March04/blog_rest_api/requirements.txt
asgiref==3.8.1
debugpy==1.8.12
Django==5.1.6
django-extensions==3.2.3
django-filter==25.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
PyJWT==2.9.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
